scripts: Move non-generated layer_base code out of python
600+ lines of normal c++ code in python strings is not maintainable.
